Added the ability to add and remove RSS feeds
RSS feeds can now be removed by swiping to delete or by pressing the
edit button.
RSS feeds can be added by pressing the "+" button and then filling in
the information and pressing the save button.
When the save button is pressed, the entered feed URL will be validated
and if the validation fails the user will be notified and the feed will
not be added.